refactor: Add screenshots to readme
