+++ src/components/About/About.css
@@ -1,5 +1,5 @@
 .About {
-    background-image: url('../../assets/images/about-wallpaper.jpg');
+    background-image: url('../../assets/images/about-wallpaper.webp');
     background-size: cover;
     background-position: center;
     padding: 2rem;
