+++ src/components/Pricing/Pricing.css
@@ -9,7 +9,7 @@
     color: #FFF;
     border-top: 5px solid #000;
     padding: 2rem;
-    background: url('../../assets/images/prices-wallpaper.png') center center/cover no-repeat;
+    background: url('../../assets/images/prices-wallpaper.webp') center center/cover no-repeat;
 }
 
 .cards-group {
